refactor: finalizada a refatoracao para utilizao unica do fluxo de post/put do agente
