+++ src/app/app.routes.ts
@@ -2,8 +2,6 @@ import { Routes } from '@angular/router';
 import { LoginComponent } from './components/back-office/login/login.component';
 import { MainPageComponent } from './components/back-office/main-page/main-page.component';
 import { AgentListComponent } from './components/back-office/agent-list/agent-list.component';
-import { AgentNewComponent } from './components/back-office/agent-new/agent-new.component';
-import { AgentEditComponent } from './components/back-office/agent-edit/agent-edit.component';
 import { AgentNewAccountComponent } from './components/back-office/agent-new-account/agent-new-account.component';
 import { AgentEditAddressComponent } from './components/back-office/agent-edit-address/agent-edit-address.component';
 import { AgentNewAddressComponent } from './components/back-office/agent-new-address/agent-new-address.component';
@@ -67,13 +65,10 @@ export const routes: Routes = [
       [
         { path: '', redirectTo: 'agent-list', pathMatch: 'full' },
         { path: 'agent-list', component: AgentListComponent, canActivate: [canActivateSupervisor] },
-        { path: 'agent-new', component: AgentNewComponent, canActivate: [canActivateSupervisor] },
-        { path: 'agent-new/:id', component: AgentNewComponent, canActivate: [canActivateSupervisor] },
         { path: 'agent-new-account/:id', component: AgentNewAccountComponent, canActivate: [canActivateSupervisor] },
         { path: 'agent-new-account/:id/:continue', component: AgentNewAccountComponent, canActivate: [canActivateSupervisor] },
         { path: 'agent-new-address/:id', component: AgentNewAddressComponent, canActivate: [canActivateSupervisor] },
         { path: 'agent-new-contact/:id', component: AgentNewContactComponent, canActivate: [canActivateSupervisor] },
-        { path: 'agent-edit/:id', component: AgentEditComponent, canActivate: [canActivateSupervisor] },
         { path: 'agent-edit-account/:id', component: AgentEditAccountComponent, canActivate: [canActivateSupervisor] },
         { path: 'agent-edit-address/:idAgent/:idAddress', component: AgentEditAddressComponent, canActivate: [canActivateSupervisor] },
         { path: 'agent-edit-contact/:idAgent/:idContact', component: AgentEditContactComponent, canActivate: [canActivateSupervisor] },
