+++ src/app/components/back-office/agent-address-list/agent-address-list.component.html
@@ -1,6 +1,6 @@
 <div class="table-header">
     <div class="header-links">
-        <a [routerLink]="['/main-page','agent-edit',agent.id]" class="return-link">Voltar</a>
+        <a [routerLink]="['/main-page','agent-register',agent.id]" class="return-link">Voltar</a>
         <a [routerLink]="['/main-page','agent-new-address',agent.id]" class="new-address-link">Novo Endereço</a>
     </div>
     <input type="text" class="search-bar" placeholder="Pesquisar...">
