+++ src/app/components/back-office/agent-contact-list/agent-contact-list.component.html
@@ -1,6 +1,6 @@
 <div class="table-header">
     <div class="header-links">
-        <a [routerLink]="['/main-page','agent-edit',agent.id]" class="return-link">Voltar</a>
+        <a [routerLink]="['/main-page','agent-register',agent.id]" class="return-link">Voltar</a>
         <a [routerLink]="['/main-page','agent-new-contact',agent.id]" class="new-contact-link">Novo Contacto</a>
     </div>
     <input type="text" class="search-bar" placeholder="Pesquisar...">
